// converter-web-client/src/app/shared/models/mappingattributes.model.ts
export class MappingAttributes {
    qmlAttribute: string;
    qmlType: string;
    qmlNested: string[];
    qmlFormat: string[];
    qmlUnit: string[];
    seisanLines: string[];
    seisanFields: string[];
    seisanColumns: string[];
    seisanFormat: string[];
    seisanUnit: string[];
    comments: string[];
}